+++ MH12Taxi/MH12Taxi/ILocationService.h
@@ -1,0 +1,7 @@
+class ILocationService
+{
+public:
+	virtual Location GetLocation(string addr) = 0;
+	virtual double CalculateDistance(string from, string to) = 0;
+};
+
