+++ MH12Taxi/MH12Taxi/Trip.h
@@ -1,0 +1,11 @@
+class Trip
+{
+public:
+	string startlocation;
+	string endlocation;
+	string startTime;
+	string endTime;
+	string paymentTime;
+	string paymentMode;
+	ICab* cab;
+};
